+++ examples/foundational/07zd-interruptible-aicoustics.py
@@ -12,7 +12,7 @@ import wave
 from dotenv import load_dotenv
 from loguru import logger
 
-from pipecat.audio.filters.aic_filter_v2 import AICFilter
+from pipecat.audio.filters.aic_filter import AICFilterV2
 from pipecat.audio.turn.smart_turn.local_smart_turn_v3 import LocalSmartTurnAnalyzerV3
 from pipecat.frames.frames import LLMRunFrame
 from pipecat.pipeline.pipeline import Pipeline
@@ -45,11 +45,12 @@ audiobuffer = AudioBufferProcessor(
 )
 
 
-def _create_aic_filter() -> AICFilter:
+def _create_aic_filter() -> AICFilterV2:
     license_key = os.getenv("AICOUSTICS_LICENSE_KEY", "")
 
-    return AICFilter(
+    return AICFilterV2(
         license_key=license_key,
+        model_id="quail-xxs-48khz",
         enhancement_level=0.5,
     )
 
@@ -62,7 +63,7 @@ transport_params = {
         lambda aic: DailyParams(
             audio_in_enabled=True,
             audio_out_enabled=True,
-            vad_analyzer=aic.create_vad_analyzer(lookback_buffer_size=6.0, sensitivity=6.0),
+            vad_analyzer=aic.create_vad_analyzer(speech_hold_duration=0.05, sensitivity=6.0),
             audio_in_filter=aic,
         )
     )(_create_aic_filter()),
@@ -70,7 +71,7 @@ transport_params = {
         lambda aic: FastAPIWebsocketParams(
             audio_in_enabled=True,
             audio_out_enabled=True,
-            vad_analyzer=aic.create_vad_analyzer(lookback_buffer_size=6.0, sensitivity=6.0),
+            vad_analyzer=aic.create_vad_analyzer(speech_hold_duration=0.05, sensitivity=6.0),
             audio_in_filter=aic,
         )
     )(_create_aic_filter()),
@@ -78,7 +79,7 @@ transport_params = {
         lambda aic: TransportParams(
             audio_in_enabled=True,
             audio_out_enabled=True,
-            vad_analyzer=aic.create_vad_analyzer(lookback_buffer_size=6.0, sensitivity=6.0),
+            vad_analyzer=aic.create_vad_analyzer(speech_hold_duration=0.05, sensitivity=6.0),
             audio_in_filter=aic,
         )
     )(_create_aic_filter()),
