+++ pyproject.toml
@@ -49,7 +49,6 @@ Changelog = "https://github.com/pipecat-ai/pipecat/blob/main/CHANGELOG.md"
 
 [project.optional-dependencies]
 aic = [ "aic-sdk>=1.2.0" ]
-#aic = [ "aic-sdk @ file:///Users/goedev/Workspace/aic/pipecat/aic_sdk-2.0.0-cp313-cp313-macosx_11_0_arm64.whl" ]
 anthropic = [ "anthropic~=0.49.0" ]
 assemblyai = [ "pipecat-ai[websockets-base]" ]
 asyncai = [ "pipecat-ai[websockets-base]" ]
